+++ src/cusolver_wrapper.cuh
@@ -17,10 +17,10 @@
 #pragma once
 
 
-void eig_hermitian_QR(int, double*, double*, double*);
-void eig_hermitian_Jacobi(int, double*, double*, double*);
-void eig_hermitian_Jacobi_batch(int, int, double*, double*, double*);
+void eig_hermitian_QR(size_t, double*, double*, double*);
+void eig_hermitian_Jacobi(size_t, double*, double*, double*);
+void eig_hermitian_Jacobi_batch(size_t, size_t, double*, double*, double*);
 void eigenvectors_symmetric_Jacobi
-(int N, double* A, double* W_cpu, double* eigenvectors_cpu);
+(size_t N, double* A, double* W_cpu, double* eigenvectors_cpu);
